+++ src/app/all-modules/all-modules.component.html
@@ -1,6 +1,6 @@
 <div class="all-modules">
   <h1>Modulhandbücher</h1>
   <mat-nav-list>
-    <a mat-list-item [href]="urlFor(m)" target='_blank' *ngFor="let m of metadata"> {{ m.title }} </a>
+    <a mat-list-item [href]="urlFor(m)" target='_blank' *ngFor="let m of modules"> {{ m.title }} </a>
   </mat-nav-list>
 </div>
