+++ src/app/create-or-update-module/callbacks/competence-callback.ts
@@ -1,5 +1,5 @@
 import { SingleEntryCallback } from './single-entry-callback'
-import { Competence } from '../../http/http.service'
+import { Competence } from '../../types/core/competence'
 
 export class CompetenceCallback extends SingleEntryCallback<Competence> {
   constructor(all: Readonly<Competence[]>, selected: Readonly<Competence[]>, attr: string, show: (m: Competence) => string) {
