+++ src/app/create-or-update-module/callbacks/global-criteria-callback.ts
@@ -1,5 +1,5 @@
 import { SingleEntryCallback } from './single-entry-callback'
-import { GlobalCriteria } from '../../http/http.service'
+import { GlobalCriteria } from '../../types/core/global-criteria'
 
 export class GlobalCriteriaCallback extends SingleEntryCallback<GlobalCriteria> {
   constructor(all: Readonly<GlobalCriteria[]>, selected: Readonly<GlobalCriteria[]>, attr: string, show: (m: GlobalCriteria) => string) {
