+++ src/app/create-or-update-module/callbacks/lecturer-callback.ts
@@ -1,5 +1,5 @@
-import { Person } from '../../http/http.service'
 import { SingleEntryCallback } from './single-entry-callback'
+import { Person } from '../../types/core/person'
 
 export class LecturerCallback extends SingleEntryCallback<Person> {
   constructor(all: Readonly<Person[]>, selected: Readonly<Person[]>, attr: string, show: (m: Person) => string) {
