+++ src/app/create-or-update-module/callbacks/module-callback.ts
@@ -1,5 +1,5 @@
-import { Module } from '../../http/http.service'
 import { SingleEntryCallback } from './single-entry-callback'
+import { Module } from '../../types/module'
 
 export class ModuleCallback extends SingleEntryCallback<Module> {
   constructor(all: Readonly<Module[]>, selected: Readonly<Module[]>, attr: string, show: (m: Module) => string) {
