+++ src/app/create-or-update-module/callbacks/prerequisites-po-callback.ts
@@ -1,5 +1,5 @@
 import { SingleEntryCallback } from './single-entry-callback'
-import { POPreview } from '../../http/http.service'
+import { POPreview } from '../../types/pos'
 
 export class PrerequisitesPoCallback extends SingleEntryCallback<POPreview> {
   constructor(all: Readonly<POPreview[]>, selected: Readonly<POPreview[]>, attr: string, show: (p: POPreview) => string) {
