+++ src/app/http/http.service.ts
@@ -1,270 +1,23 @@
 import { Injectable } from '@angular/core'
 import { HttpClient } from '@angular/common/http'
 import { map, Observable, zip } from 'rxjs'
-
-export interface Label {
-  deLabel: string
-  enLabel: string
-}
-
-export interface Workload {
-  lecture: number,
-  seminar: number,
-  practical: number,
-  exercise: number,
-  projectSupervision: number,
-  projectWork: number,
-  selfStudy: number,
-  total: number
-}
-
-export interface Participants {
-  min: number
-  max: number
-}
-
-export interface Parent {
-  kind: 'parent'
-  children: string[]
-}
-
-export interface Child {
-  kind: 'child'
-  parent: string
-}
-
-export type ModuleRelation = Parent | Child
-
-export interface AssessmentMethodEntry {
-  method: string,
-  percentage?: number,
-  precondition: string[]
-}
-
-export interface AssessmentMethods {
-  mandatory: AssessmentMethodEntry[],
-  optional: AssessmentMethodEntry[]
-}
-
-export interface PrerequisiteEntry {
-  text: string,
-  modules: string[],
-  pos: string[]
-}
-
-export interface PrerequisitesOutput {
-  recommended?: PrerequisiteEntry,
-  required?: PrerequisiteEntry
-}
-
-export interface POMandatory {
-  po: string,
-  recommendedSemester: number[],
-  recommendedSemesterPartTime: number[]
-}
-
-export interface POOptional {
-  po: string,
-  instanceOf: string,
-  partOfCatalog: boolean,
-  recommendedSemester: number[]
-}
-
-export interface POs {
-  mandatory: POMandatory[],
-  optional: POOptional[]
-}
-
-export interface Metadata {
-  id: string,
-  title: string,
-  abbrev: string,
-  moduleType: string,
-  ects: number,
-  language: string,
-  duration: number,
-  season: string,
-  workload: Workload,
-  status: string,
-  location: string,
-  participants?: Participants,
-  moduleRelation?: ModuleRelation,
-  moduleManagement: string[],
-  lecturers: string[],
-  assessmentMethods: AssessmentMethods,
-  prerequisites: PrerequisitesOutput,
-  po: POs,
-  competences: string[],
-  globalCriteria: string[],
-  taughtWith: string[]
-}
-
-export interface MetadataPreview {
-  id: string
-  title: string
-  abbrev: string
-}
-
-export interface Location {
-  abbrev: string
-  deLabel: string
-  enLabel: string
-}
-
-export interface Language {
-  abbrev: string
-  deLabel: string
-  enLabel: string
-}
-
-export interface Status {
-  abbrev: string
-  deLabel: string
-  enLabel: string
-}
-
-export interface AssessmentMethod {
-  abbrev: string
-  deLabel: string
-  enLabel: string
-}
-
-export interface ModuleType {
-  abbrev: string
-  deLabel: string
-  enLabel: string
-}
-
-export interface Season {
-  abbrev: string
-  deLabel: string
-  enLabel: string
-}
-
-export interface Faculty {
-  abbrev: string
-  deLabel: string
-  enLabel: string
-}
-
-export interface PersonStatus {
-  deLabel: string
-  enLabel: string
-}
-
-export interface SinglePerson {
-  id: string
-  lastname: string
-  firstname: string
-  title: string
-  faculties: Faculty[]
-  abbreviation: string
-  status: PersonStatus
-  kind: 'single'
-}
-
-export interface UnknownPerson {
-  id: string
-  title: string
-  kind: 'unknown'
-}
-
-export interface GroupPerson {
-  id: string
-  title: string
-  kind: 'group'
-}
-
-export type Person = SinglePerson | UnknownPerson | GroupPerson
-
-export interface StudyFormType {
-  abbrev: string
-  deLabel: string
-  enLabel: string
-}
-
-export interface PO {
-  abbrev: string,
-  version: number,
-  date: Date,
-  dateFrom: Date,
-  dateTo?: Date,
-  modificationDates: Date[],
-  program: string
-}
-
-export interface POPreview {
-  id: string
-  label: string
-  abbrev: string
-}
-
-export interface Grade {
-  abbrev: string
-  deLabel: string
-  enLabel: string
-  deDesc: string
-  enDesc: string
-}
-
-export interface FocusArea {
-  abbrev: string
-  program: string
-  deLabel: string
-  enLabel: string
-  deDesc: string
-  enDesc: string
-}
-
-export interface GlobalCriteria {
-  abbrev: string
-  deLabel: string
-  enLabel: string
-  deDesc: string
-  enDesc: string
-}
-
-export interface RestrictedAdmission {
-  value: boolean
-  deReason: string
-  enReason: string
-}
-
-export interface StudyProgram {
-  abbrev: string,
-  deLabel: string,
-  enLabel: string,
-  internalAbbreviation: string,
-  externalAbbreviation: string,
-  deUrl: string,
-  enUrl: string,
-  grade: string,
-  programDirector: string,
-  accreditationUntil: Date,
-  restrictedAdmission: RestrictedAdmission,
-  studyForm: string[],
-  language: string[],
-  seasons: string[],
-  campus: string[],
-  deDescription: string,
-  deNote: string,
-  enDescription: string,
-  enNote: string
-}
-
-export interface Competence {
-  abbrev: string
-  deLabel: string
-  enLabel: string
-  deDesc: string
-  enDesc: string
-}
-
-export interface Module {
-  id: string
-  title: string
-  abbrev: string
-}
+import { Metadata } from '../types/metadata'
+import { Location } from '../types/core/location'
+import { Language } from '../types/core/language'
+import { Status } from '../types/core/status'
+import { AssessmentMethod } from '../types/core/assessment-method'
+import { ModuleType } from '../types/core/module-type'
+import { Season } from '../types/core/season'
+import { Faculty } from '../types/core/faculty'
+import { Person } from '../types/core/person'
+import { StudyFormType } from '../types/core/study-form-type'
+import { PO } from '../types/core/po'
+import { Grade } from '../types/core/grade'
+import { FocusArea } from '../types/core/focus-area'
+import { GlobalCriteria } from '../types/core/global-criteria'
+import { StudyProgram } from '../types/core/study-program'
+import { Competence } from '../types/core/competence'
+import { Module } from '../types/module'
 
 @Injectable({
   providedIn: 'root'
