+++ src/app/types/assessment-methods.ts
@@ -1,0 +1,10 @@
+export interface AssessmentMethods {
+  mandatory: AssessmentMethodEntry[],
+  optional: AssessmentMethodEntry[]
+}
+
+export interface AssessmentMethodEntry {
+  method: string,
+  percentage?: number,
+  precondition: string[]
+}
