+++ src/app/types/core/assessment-method.ts
@@ -1,0 +1,5 @@
+export interface AssessmentMethod {
+  abbrev: string
+  deLabel: string
+  enLabel: string
+}
