+++ src/app/types/core/competence.ts
@@ -1,0 +1,7 @@
+export interface Competence {
+  abbrev: string
+  deLabel: string
+  enLabel: string
+  deDesc: string
+  enDesc: string
+}
