+++ src/app/types/core/faculty.ts
@@ -1,0 +1,5 @@
+export interface Faculty {
+  abbrev: string
+  deLabel: string
+  enLabel: string
+}
