+++ src/app/types/core/focus-area.ts
@@ -1,0 +1,8 @@
+export interface FocusArea {
+  abbrev: string
+  program: string
+  deLabel: string
+  enLabel: string
+  deDesc: string
+  enDesc: string
+}
