+++ src/app/types/core/global-criteria.ts
@@ -1,0 +1,7 @@
+export interface GlobalCriteria {
+  abbrev: string
+  deLabel: string
+  enLabel: string
+  deDesc: string
+  enDesc: string
+}
