+++ src/app/types/core/grade.ts
@@ -1,0 +1,7 @@
+export interface Grade {
+  abbrev: string
+  deLabel: string
+  enLabel: string
+  deDesc: string
+  enDesc: string
+}
