+++ src/app/types/core/label.ts
@@ -1,0 +1,4 @@
+export interface Label {
+  deLabel: string
+  enLabel: string
+}
