+++ src/app/types/core/language.ts
@@ -1,0 +1,5 @@
+export interface Language {
+  abbrev: string
+  deLabel: string
+  enLabel: string
+}
