+++ src/app/types/core/location.ts
@@ -1,0 +1,5 @@
+export interface Location {
+  abbrev: string
+  deLabel: string
+  enLabel: string
+}
