+++ src/app/types/core/module-type.ts
@@ -1,0 +1,5 @@
+export interface ModuleType {
+  abbrev: string
+  deLabel: string
+  enLabel: string
+}
