+++ src/app/types/core/po.ts
@@ -1,0 +1,9 @@
+export interface PO {
+  abbrev: string,
+  version: number,
+  date: Date,
+  dateFrom: Date,
+  dateTo?: Date,
+  modificationDates: Date[],
+  program: string
+}
