+++ src/app/types/core/season.ts
@@ -1,0 +1,5 @@
+export interface Season {
+  abbrev: string
+  deLabel: string
+  enLabel: string
+}
