+++ src/app/types/core/status.ts
@@ -1,0 +1,5 @@
+export interface Status {
+  abbrev: string
+  deLabel: string
+  enLabel: string
+}
