+++ src/app/types/core/study-form-type.ts
@@ -1,0 +1,5 @@
+export interface StudyFormType {
+  abbrev: string
+  deLabel: string
+  enLabel: string
+}
