+++ src/app/types/module-relation.ts
@@ -1,0 +1,11 @@
+export type ModuleRelation = Parent | Child
+
+export interface Parent {
+  kind: 'parent'
+  children: string[]
+}
+
+export interface Child {
+  kind: 'child'
+  parent: string
+}
