+++ src/app/types/module.ts
@@ -1,0 +1,5 @@
+export interface Module {
+  id: string
+  title: string
+  abbrev: string
+}
