+++ src/app/types/participants.ts
@@ -1,0 +1,4 @@
+export interface Participants {
+  min: number
+  max: number
+}
