+++ src/app/types/prerequisites.ts
@@ -1,0 +1,10 @@
+export interface PrerequisitesOutput {
+  recommended?: PrerequisiteEntry,
+  required?: PrerequisiteEntry
+}
+
+export interface PrerequisiteEntry {
+  text: string,
+  modules: string[],
+  pos: string[]
+}
