+++ src/app/types/workload.ts
@@ -1,0 +1,10 @@
+export interface Workload {
+  lecture: number,
+  seminar: number,
+  practical: number,
+  exercise: number,
+  projectSupervision: number,
+  projectWork: number,
+  selfStudy: number,
+  total: number
+}
